+++ frontend/src/app/pages/exercises/exercise/shared/trainer-map-editor/trainer-map-editor.component.ts
@@ -6,11 +6,7 @@ import type {
     UUID,
     VehicleTemplate,
 } from 'digital-fuesim-manv-shared';
-import {
-    TransferPoint,
-    Viewport,
-    SimulatedRegion,
-} from 'digital-fuesim-manv-shared';
+import { TransferPoint, Viewport } from 'digital-fuesim-manv-shared';
 import type { AppState } from 'src/app/state/app.state';
 import {
     selectMapImagesTemplates,
@@ -21,6 +17,7 @@ import { DragElementService } from '../core/drag-element.service';
 import { TransferLinesService } from '../core/transfer-lines.service';
 import { openCreateImageTemplateModal } from '../editor-panel/create-image-template-modal/open-create-image-template-modal';
 import { openEditImageTemplateModal } from '../editor-panel/edit-image-template-modal/open-edit-image-template-modal';
+import { simulatedRegionDragTemplates } from '../editor-panel/templates/simulated-region';
 
 @Component({
     selector: 'app-trainer-map-editor',
@@ -57,9 +54,7 @@ export class TrainerMapEditorComponent {
         private readonly ngbModalService: NgbModal
     ) {}
 
-    public readonly simulatedRegionTemplate = {
-        image: SimulatedRegion.image,
-    };
+    public readonly simulatedRegionTemplates = simulatedRegionDragTemplates;
 
     public readonly viewportTemplate = {
         image: Viewport.image,